Move back to double quotes
Don't know why I felt that JS/TS should only use single quotes,
and it just adds more changes to the tree for what seems like
an arbitrary reason. So moving back to double quotes.

+++ eslint.config.mjs
@@ -27,7 +27,7 @@ export default defineConfig([
         rules: {
             "@typescript-eslint/no-explicit-any": "off",
             'react/prop-types': 'off',
-            '@stylistic/quotes': ['error', 'single'],
+            '@stylistic/quotes': ['error', 'double'],
             '@stylistic/no-extra-semi': 'error',
             '@stylistic/semi': ['error', 'always'],
             '@stylistic/curly-newline': 'error',
